wave plane test

+++ Code/WMFramework/Graphics/Geometry/WMMesh.h
@@ -16,7 +16,7 @@ namespace WildMini::Graphics::Geometry
     struct WMMesh
     {
         std::vector<WildMini::Graphics::Geometry::WMVertex> vertices;
-        std::vector<size_t> indices;
+        std::vector<uint32_t> indices;
         WildMini::WMSharedPtr<WildMini::Graphics::WMGPUBuffer> vertexBuffer;
         WildMini::WMSharedPtr<WildMini::Graphics::WMGPUBuffer> indexBuffer;
